Made initial art files and test map
Made initial art files and test map. These do not fall under the main project license and thus I hold the exclusive copyrights for now. I am planning to add a proper licensing to these such as a creative commons license or similar. Note that none of this art is final yet, it might change in the future!